[signals][breaking] Reject the promise when the process exits due to a signal w/out code 0
When a process exits because of a signal like SIGKILL, SIGTERM, or SIGINT, this makes `spawnAsync` reject the promise (i.e. throw an async error) instead of returning normally. This is because the process did not exit with code 0 -- Node sets the code to `null` and specifies the signal instead.

Test plan: Updated unit test to check for this case using a small program that kills itself using a signal.

Closes #9