[tests] Add test that extra listeners (ex: error, close) are run
Our error and close event listeners call `removeAllListeners()` but we want to make sure that other externally added error and close listeners run when there is one of those events.

Note: this PR doesn't address running extra close listeners when there is an error event -- just that extra error listeners are run

Test plan: added a unit test to verify that extra error listeners are run when there is an error.