[exit] Add "ignoreStdio" option to listen to the "exit" event instead of "close"
When a subprocess exits, Node will fire the "exit" event even if the subprocess's stdio streams are open. Usually this doesn't matter as the "close" event is fired soon after, but in some cases -- especially when sharing or piping I/O streams -- the streams may stay open after a subprocess exits.

To listen to the "exit" event, this commit introduces a new option named `ignoreStdio`. It is named this way because it does not wait for the stdio streams to close and also does not read data from them. This is a deliberate decision since a very small fraction of the time, the "exit" event may fire before we can collect all data from stdout. This would be very hard to debug and non-deterministic, so when listening to the "exit" event we never read from stdout/stderr and just return empty strings.

Test plan: Added a test that hangs when `ignoreStdio` is false/omitted that verifies the promise resolves even if stdout is open.